// lib/platform-adapters/soundexchange/types.ts
export interface SoundExchangeRow {
  track_title: string;
  artist_name: string;
  featured_artist: string;
  isrc: string;
  album_title: string;
  label: string;
  duration: string;
  release_date: string;
  p_line_owner: string;
  rights_owner: string;
  featured_performer: string;
  non_featured_performer: string;
  producer: string;
  country: string;
  notes: string;
}